Do not return error from -i when no interfaces are configured
The -i and -l options have an optional pattern argument, and it makes
sense to return an error when the pattern does not match.

However, when no pattern is specified and no interface is configured an
error is returned as well.

This makes it problematic to detect if -i option is supported - an error
is returned both when the option is not supported by other resolvconf
implementation, and when there are no interfaces configured in
openresolv.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>