libc: provide a toggle for moving vs writing to resolv.conf
Creating a temp file and moving it to resolv.conf as we did in the last
release has created too many issues.
As such the default again is to write directly to it.

To try and appease all user needs, a configurable `resolv_conf_mv`
can now be set to a truthy value for those that want this still.

Fixes #19, partially reverts #6.