Add a new exlcude option to exclude a whole file on match
Has a similar syntax to replace.
While here add a new -L option to list the resolv.conf files
we have and apply post-processing so we can easily test.

Also while here implement a poor mans cat.
This means our needs and avoids a fork. This is fine as
cat may exist in /usr on some platforms which may not be available
in early boot for us and the files we do read shouldn't be many
lines long.