resolvectl: initial suport for it
resolvconf as supplied by systemd-resolved doesn't work with
different programs or protocols suppling dns for the same interface.

This new subscriber allows openresolv to be installed as resolvconf
and still allows systemd-resolved to work fine.

Any resolvconf entries that cannot be matched to an interface
are emitted as a warning on update - this is a systemd-resolved
limitation. If you are beholden to it, use something else!

You may want to consider setting resolv_conf=/dev/null in
/etc/resolvconf.conf to avoid resolvconf stamping on the
systemd-resolved /etc/resolv.conf file.