resolvectl: Add support for it again
But really, you shouldn't be using this long term.
Once systemd-resolved get a release supports DNS delegates, the
only use for this new subscriber is to show your DNS setup
per interface how systemd-resolved likes it, as opposed to
DNS delegates which is how every other resolver likes it.

Fixes #34.